@@ -18,19 +18,19 @@
 		}
 	],
 	"require": {
-		"php": "^7.1",
+		"php": "^7.4 || ^8.0",
 		"ext-json": "*",
 		"nette/application": "~2.4|~3.0.0|~3.1.0",
-		"nette/forms": "~2.4|~3.0.0",
-		"contributte/recaptcha": "3.1.* | 3.2.* | 3.3.*",
+		"nette/forms": "~2.4|~3.0.0|~3.1.0",
+		"contributte/recaptcha": "^3.1",
 		"nette/di": "~2.4|~3.0.0",
-		"nette/utils": "~2.4|~3.0.0|~3.1.0",
+		"nette/utils": "~2.4|~3.0.0|~3.1.0|~3.2.0",
 		"pd/utils": "^1.0"
 	},
 	"require-dev": {
 		"nette/tester": "~2.3.2",
 		"mockery/mockery": "^1.0",
-		"pd/coding-standard": "1.21.*",
+		"pd/coding-standard": "1.27.*",
 		"phpstan/phpstan": "0.12.77"
 	},
 	"autoload": {
