@@ -9,10 +9,8 @@ require __DIR__ . '/../../bootstrap.php';
  */
 final class RuleOptionsFactoryTest extends \Tester\TestCase
 {
-	/**
-	 * @var \Pd\Forms\RuleOptionsFactory
-	 */
-	private $ruleOptionsFactory;
+
+	private \Pd\Forms\RuleOptionsFactory $ruleOptionsFactory;
 
 
 	public function testCreateOptions(): void
@@ -39,17 +37,21 @@ final class RuleOptionsFactoryTest extends \Tester\TestCase
 
 		if (\interface_exists(\Nette\Schema\Schema::class)) { //nette 3.0
 			$translator = new class() implements \Nette\Localization\ITranslator {
+
 				function translate($message, ...$parameters): string
 				{
 					return '';
 				}
-			};
+
+};
 		} else {
 			$translator = new class() implements \Nette\Localization\ITranslator {
+
 				public function translate($message, $count = NULL)
 				{
 				}
-			};
+
+};
 		}
 
 		$this->ruleOptionsFactory = new \Pd\Forms\RuleOptionsFactory($translator);
